test-requirements.txt
# Testing requirements
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
pytest-mock==3.12.0
fastapi==0.104.1
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-saml==1.15.0
sqlalchemy==2.0.23
pyotp==2.9.0
